Improve choice of tycon pretty names during signature match
Refinement of 27cb830

Consider the following program:

    signature S = sig
      type t
      datatype u = U
      sharing type t = u
    end
    structure S : S = struct
    end

Previously, MLton generated the following error messages:

    Error: z.sml 6.15-6.15.
      Type in signature but not in structure: t.
        signature: type t = _sig.t
        spec at:   z.sml 2.8-2.8
        spec at:   z.sml 4.3-4.20
    Error: z.sml 6.15-6.15.
      Type in signature but not in structure: u.
        signature: datatype u = U
        spec at:   z.sml 3.12-3.12
        spec at:   z.sml 4.3-4.20

Note that the first message has the vacuous `type t = _sig.t`.  This
is because both `t` and `u` are names for the same (unrealized) type
constructor and `t` is chosen as the representative name.  The
incongruity is that the `u` was carefuly chosen as the flexible tycon
to be realized.

Now, MLton generates the following error messages:

    Error: z.sml 6.15-6.15.
      Type in signature but not in structure: t.
        signature: type t = _sig.u
        spec at:   z.sml 2.8-2.8
        spec at:   z.sml 4.3-4.20
    Error: z.sml 6.15-6.15.
      Type in signature but not in structure: u.
        signature: datatype u = U
        spec at:   z.sml 3.12-3.12
        spec at:   z.sml 4.3-4.20

Rather than processing the realized signature with lowest priority,
the flexible tycon map is processed with lowest priority.