Improve "escaping type" diagnostics
Consider the following program:

    fun app (z, f) = f z
    val _ =
       app
       (let
           datatype t = T
           datatype 'a u = U of 'a
        in
           (U T, fn U T => ())
        end)

Previously, MLton generated the following error messages:

    Error: z.sml 2.1-9.8.
      Type escapes the scope of its definition: u.
        defn at: z.sml 6.20-6.20
        in: val _ = app let datatype t = T da  ...  T, fn U T => ()) end
    Error: z.sml 2.1-9.8.
      Type escapes the scope of its definition: t.
        defn at: z.sml 5.17-5.17
        in: val _ = app let datatype t = T da  ...  T, fn U T => ()) end

Note that the region of both errors is the declaration, rather than
the let expression.

Now, MLton generates the following error message:

    Error: z.sml 4.5-9.7.
      Type of let has escaping local types: t, u.
        type: [t u] * ([t u] -> _)
        in: let datatype t = T datatype 'a u   ...  T, fn U T => ()) end

The implementation is a modification of TypeEnv.minType; rather than
invoke an error handler when a tycon is used before it is defined, a
layout object is returned highlighting the out-of-scope tycons.  The
implementation is a little more sophisticated than necessary for just
this diagnostic in order to prepare for a similar change for minTime
as used by unify.