Properly separate AST Tyvar from CoreML Tyvar
The same Atoms.Tyvar structure had been used for both AST and
CoreML/XML ILs.  One had to be careful which kind of equality to use
at various points.  Also, the old Tyvar.newNoname incurred an
Int.toString conversion to generate the unique name; the new
Tyvar.newNoname generates the printName lazily (as with other IDs).

The elaborate of Ast.Tyvar to Atoms.Tyvar is handled by a simple
TyvarEnv.  Since the scope of a tyvar is at most a single Core dec or
a single Module spec, the tyvar environment is not integrated with the
main ElaborateEnv datatype.

Currently, one can get duplicate "undefined type variable" errors for
a program, due to errors being generated by both Scope and
ElaborateCore.