Simplify ElaborateEnv.Equality type and operations
The previous implementation of ElaborateEnv.Equality.t was of the
form:

    datatype maybe =
       Known of bool
     | Unknown of {whenKnown: (bool -> bool) list ref}
    datatype t =
       False
     | Lazy of unit -> t
     | Maybe of maybe ref
     | True

The "admits equality" property in SML is complicated by two features:
type inference and "maximize equality" of type environments (i.e.,
datatype declarations).  The previous implementation conflates those
two features: the `Lazy of unit -> t` variant is used for type
constructors applied to arguments to accomodate change to type
constructor equality status during "maximize equality".  The `Maybe of
maybe ref` variant is used for type inference; unification variables
are introduced with an Unknown (except for unification variables
instantiating type schemes with equality type variables), with
`whenKnown` constraints added during unifications.

However, this implementation is complicated and many type constructors
have `AdmitsEquality.Sometimes` status and so induce `Lazy` equality
status.

The new implementation of ElaborateEnv.Equality.t is of the form

    datatype t = False | True | Unknown

with Type.t carrying an `Equality.t ref`.  Another motivation is to
treat equality in a manner similar to that of time: during
unification, only unification of an Unknown type requires a time and
equality check.  The equality propert of a type is still computed
lazily with respect to type inference and unification variables --- if
a typeis required to be an equality type (when unified with an Unknown
that is an equality type) and its root node has equality Unknown, then
the type is explored to determine if the type can be made an equality
type (updating the `Equality.t ref` fields of nodes as appropriate).
If the type cannot be made and equality type, then a description of
the violation is returned.  Again, this makes the `checkTime` and
`checkEquality` functions very similar.  One slight complication of
`checkEquality` is that one must be careful to not descend into Con
types before checking the equality status of the type constructor,
because an AdmitsEquality.Always type constructor must not force type
arguments to be equality types.

Removing the `Lazy of unit -> t` variant makes "maximize equality"
slightly more complicated.  In particular, it is necessary to
"rebuild" the argument type of data constructors and the definition
type of `withtype` clauses in order to propagate the change in type
constructor equality statu.  However, this overhead only occurs during
"maximize equality"; everywhere else, a type constructor has immutable
equality status.