Avoid allocating a TypeEnv.hom in ElaborateCore.elaborateDec
The `layoutPrettyType` function is only invoked on an error/warning,
which should be rare.  Rather than using `Type.makeLayoutPretty`,
which prepares a TypeEnv.hom, use `Type.layoutPretty` which delays the
TypeEnv.hom until called.  The downside is that distinct errors do not
get to share layout objects for displayed types.