Revise ElaborateEnv.InterfaceEnv.makeLayoutPrettyFlexTycon
In order to avoid shadowing of structure tycons when displaying
signature tycons, signature types are always displayed as
"_sig.S1.S2.t" relative to the top-level signature.  However, it is
sometimes necessary to display signature types during elaboration of a
signature.  Hence, we generate types from the interface environment
plus the "current" interface with a nesting.