Restore behavior of `-keep-def-use`
The schemes registered for def-use information are not always true
schemes.  For example, consider the following program:

    fun f (x: 'a) (y: 'b) =
       let
          fun g (z: 'c) = (z, y, x)
       in
          g ()
       end

The "scheme" registered for `f` and `g` are actually mono-types: the
unification types created for `f` and `g` at the point where they are
initially added to the environment.  During inference, these
unification types are unfied with function type and, at
generalization, the remaining unification types are replaced by type
variables.  Hence, the "scheme" that remains for `f` and `g` are the
mono-types with free type variables.  (Similarly, the "scheme"
registered for `x`, `y`, and `z` are mono-types with free type
variables.)

To yield reasonable def-use extra information, these "schemes" are
pretty-printed with fresh local names for each distinct type
variable.  This makes each scheme look reasonable in isolation, but the
total information is not globally consistent.  For example, the
def-use information for the above program is:

    variable f z.sml 2.5 "'a -> 'b -> unit * 'b * 'a"
    variable x z.sml 2.8 "'a"
        z.sml 4.30
    variable y z.sml 2.16 "'a"
        z.sml 4.27
    variable g z.sml 4.11 "'a -> 'a * 'b * 'c"
        z.sml 6.7
    variable z z.sml 4.14 "'a"
        z.sml 4.24

Note that each of `x`, `y`, and `z` reports the type `'a`, despite the
fact that they must be distinct type variables.

This is probably better than the previous behavior (on this
branch) which would fall-thru to the default `Tyvar.layout` (with
gen-sym names):

    variable f z.sml 2.5 "'a_0 -> 'b_0 -> unit * 'b_0 * 'a_0"
    variable x z.sml 2.8 "'a_0"
        z.sml 4.30
    variable y z.sml 2.16 "'b_0"
        z.sml 4.27
    variable g z.sml 4.11 "'c_0 -> 'c_0 * 'b_0 * 'a_0"
        z.sml 6.7
    variable z z.sml 4.14 "'c_0"
        z.sml 4.24

While this information is globally consistent, in a full program,
with many inferred type variables, most polymorphic values have
schemes with `'a_2342`-style type variables, which are difficult
to read.

"Pretty" and "globally consistent" information would require some
effort.  In any case, this restores the (master branch) behavior
of `-keep-def-use`.