Fix "Interface.FlexibleTycon.realization" internal compiler error
Consider the following program:

    signature T =
       sig
          type u
          type t
       end
    signature S =
       sig
          type u
          include T where type t = u -> u
       end
    structure S: S =
       struct
          type u = unit
          type t = u -> u
       end
    functor F (S: S): S = struct open S end

Previously, MLton generated an ICE:

    Error: z.sml 9.15-9.37.
      Duplicate type specification: u.
        spec at: z.sml 8.12-8.12
        spec at: z.sml 3.12-3.12
    unhandled exception: Fail: Interface.FlexibleTycon.realization

This occurs at the `structure S: S = ...` matching (and would occur
again at the creation of the dummy structure for the functor argument
and again at the "functor F (S: S): S = ...` result matching.  The
issue is that the shadowed `type u` in `signature S` is a flexible
tycon, but because it is shadowed, it is not reported as a flexible
tycon nor is it realized.  Nonetheless, the `type t` is realized as
`u -> u` for the shadowed `u`, which requires obtaining the
realization of the shadowed `u` to check the realized type of `t`.

Now, MLton generates the following error messages:

    Error: z.sml 9.15-9.37.
      Duplicate type specification: u.
        spec at: z.sml 8.12-8.12
        spec at: z.sml 3.12-3.12
    Error: z.sml 11.14-11.14.
      Type in structure disagrees with signature (type definition): t.
        structure: type t = [unit] -> [unit]
        defn at: z.sml 14.12-14.12
        signature: type t = [?.??.S.u] -> [?.??.S.u]
        spec at: z.sml 4.12-4.12
        spec at: z.sml 9.17-9.37
    Error: z.sml 16.19-16.19.
      Type in structure disagrees with signature (type definition): t.
        structure: type t = [?.??.S.u] -> [?.??.S.u]
        defn at: z.sml 4.12-4.12
        signature: type t = [?.??.S.u] -> [?.??.S.u]
        spec at: z.sml 4.12-4.12
        spec at: z.sml 9.17-9.37

It is difficult to give a good pretty name for this realization of a
shadowed flexible tycon, but it is also a rather obscure kind of error
and the "?.??." prefix seems to convey the difficulty.