Add `Array_uninit: 'a array * SeqInt.int -> unit` primitive
The Array_uninit primitive sets all objptrs in the element at the
given index to a bogus non-objptr value (0wx1).  One motivation for
this primitive is to support space-efficient polymorphic resizeable
arrays.  When shrinking a resizeable array, we would like to "NULL"
out the elements that are no longer part of the logical array, in
order to avoid a (logical) space leak.