Add `Array_uninitIsNop: 'a array -> bool` primitive
The Array_uninitIsNop primitive answers if the Array_uninit primitive
applied to the same array would be a nop (i.e., if the array has no
objptrs in the elements).  This can be used to skip a
bulk-Array_uninit loop when it is known that the Array_uninit
operations would be nops.

The primitive is translated to a constant by toRssa.

It would be possible to introduce SSA and SSA2 optimizations that
simplified Array_uninitIsNop and Array_uninit when it is clear that
the array will never have objptrs in the elements.  In particular,
while an `(int64 * int64) array` (in SSA) might become an `int64
array` (due to Useless), the reverse can never happen.  Similarly,
while an `((int64 * int64)) array` (in SSA2) might become an `(int64 *
int64) array` (due to DeepFlatten), the reverse can never happen.

Also, RSSA shrink does not seem to simplify the

    switch {test = 0x0: Word32, default = None, cases = ((0x0, L_1277), (0x1, L_1276))}

that results from replacing Array_uninitIsNop with a constant.