Add `Array_allocRaw: SeqIndex.int -> 'a array` primitive
The `Array_allocRaw` primitive allocates an array, but with a header
that indicates that the array has no objptrs.  A related
`Array_toArray` primitive updates the header of an `Array_allocRaw`
primitive to reveal the objptrs.  Between the allocation and the cast,
all elements of the array must be uninitialized (setting all objptrs
in elements to a bogus non-objptr value (0wx1)).

One motiviation for this primitive is that, in a parallel setting, the
uninitialization of an array can be a sequential bottleneck.  The
`Array_allocRaw` is a constant time operation and the subsequent
`Array_uninit` operations can be performed in parallel.

Note that it is not safe to write non-bogus objptrs into a raw array.
Because the raw array's header indicates that the array has no
objptrs, non-bogus objptrs written to the array will not be traced or
updated during a garbage collection.

In the future, we may consider eliminating `Array_alloc` (the non-raw
array allocation) and moving the uninitialization loop from the
runtime to SML.  To do so, though, we should ensure that arrays that
do not need uninitialization are optimized early in the optimization
pipeline.