Implement alternate canonicalization strategies in CommonSubexp
As part of performing common subexpression elimination, CommonSubexp
canonicalizes all Exp.t-s, including putting arguments of commutative
arithmetic primitives in a canonical order.  There is a note in
common-subexp.fun regarding this canonicalizing of commutative
arguments:

                  fun canon2 () =
                     let
                        val a0 = arg 0
                        val a1 = arg 1
                     in
                        (* What we really want is a total orderning on
                         * variables.  Since we don't have one, we just use
                         * the total ordering on hashes, which means that
                         * we may miss a few cse's but we won't be wrong.
                         *)
                        if Var.hash a0 <= Var.hash a1
                           then (a0, a1)
                        else (a1, a0)
                     end

Note that the only cse-s that a lack of a total ordering would miss
are "a + b" and "b + a", where the variables "a" and "b" both had the
same hash.

On the other hand, the order of arguments to commutative arithmetic
arguments is sensitive to variable hashes and variable hashes are
created by an unseeded pseudo-random number generator.  This
pseudo-random number generator is used create hashes for all classes
of identifiers, so while variable hashes are deterministic, they are
sensitive to the number of identifiers created before the variable is
created.  Hence, the creation of more or less identifiers (for
example, when resuming compilation from a saved .sxml program) the
assignment of hashes to variable identifiers, which might swap the
arguments of commutative arithmetic primitives.  Indeed, the only
difference between the .ssa when compiling the original program and
the .ssa when resuming compilation from a corresponding saved .sxml
file is the order of some commutative arithmetic primitives.

One would hope that the codegens would not be sensitive to the
ordering of arguments in commutative arithmetic primitives.  Indeed,
the native codegens do some analysis to pick a "good" order in which
to handle the arguments of commutative arithmetic primitives.
However, when there isn't anything to break the tie when looking for a
"good" order, the generated code follows the order of the arguments in
the primitive.  We've seen situations where a small code permutation,
which results in the swapping of some commutative arithmetic primitive
arguments leads to non-trivial run-time differences.

This commit introduces four different canonicalization strategies for
arguments to commutative arithmetic primitives.  "ascHash" and
"descHash" correspond to the strategies that put the arguments in
order of ascending or descending variable hashes, respectively.
"ascIndex" and "descIndex" correspond to strategies that put the
arguments in order of ascending or descending variable indices,
respectively.  Variable indices are set according to a counter and a
variable's index is set when it's defining occurrence is encountered
during the walk of the dominator tree.  This doesn't guarantee that a
variable's index is always the same under different compiles; a
variable's index is sensitive to the number of variables set before
it, which will depend upon the particular program and previous program
transformations.  However, this indexing strategy does have the
advantage that the relative order of two variables will always
correspond to the relative scope of the two variables, which will tend
to be stable across compiles, so long as the variables in question are
not the subject of a serious optimization.

Here are the results:

    MLton0 -- ~/devel/mlton/builds/20171013.233239-g5513092e3/bin/mlton
    MLton1 -- ~/devel/mlton/mlton.git/build/bin/mlton -cse-canon ascHash
    MLton2 -- ~/devel/mlton/mlton.git/build/bin/mlton -cse-canon descHash
    MLton3 -- ~/devel/mlton/mlton.git/build/bin/mlton -cse-canon ascIndex
    MLton4 -- ~/devel/mlton/mlton.git/build/bin/mlton -cse-canon descIndex
    run time ratio
    benchmark         MLton1 MLton2 MLton3 MLton4
    DLXSimulator        0.99   0.99   1.00   1.00
    barnes-hut          1.00   1.00   1.05   1.03
    boyer               0.99   0.99   0.99   0.99
    checksum            1.00   1.01   1.00   0.90
    count-graphs        1.01   1.01   1.01   1.00
    even-odd            1.00   1.00   1.00   1.00
    fft                 0.99   0.99   0.98   1.00
    fib                 1.00   1.00   1.01   1.00
    flat-array          1.00   1.00   0.99   0.99
    hamlet              0.98   0.98   0.99   1.01
    imp-for             0.99   0.99   1.00   1.00
    knuth-bendix        1.03   0.99   1.02   0.99
    lexgen              0.97   0.94   0.95   0.93
    life                0.99   0.94   0.97   0.95
    logic               1.00   1.01   1.03   1.04
    mandelbrot          1.00   1.10   1.10   1.00
    matrix-multiply     0.99   1.04   1.04   1.00
    md5                 1.00   1.01   1.00   1.00
    merge               1.00   1.00   1.01   1.00
    mlyacc              1.01   1.01   1.00   1.01
    model-elimination   1.02   1.01   1.01   1.00
    mpuz                0.99   0.99   1.00   1.00
    nucleic             1.00   0.98   0.98   0.98
    output1             1.00   1.00   1.00   1.00
    peek                1.00   0.99   1.00   1.00
    psdes-random        0.99   0.84   0.99   0.85
    ratio-regions       1.00   1.00   1.00   1.00
    ray                 0.99   1.01   1.01   1.01
    raytrace            1.00   1.00   0.95   0.94
    simple              0.99   0.97   1.01   0.99
    smith-normal-form   1.01   1.01   1.00   1.01
    string-concat       1.00   1.00   1.00   1.00
    tailfib             1.00   1.00   1.00   1.00
    tak                 1.00   1.02   1.00   1.00
    tensor              1.00   1.00   1.00   1.00
    tsp                 1.00   0.94   1.00   1.00
    tyan                1.01   1.00   1.00   1.00
    vector-rev          0.99   1.02   0.99   0.99
    vector32-concat     1.00   1.00   1.00   1.00
    vector64-concat     0.99   1.00   1.00   1.00
    vliw                1.00   1.00   0.99   0.99
    wc-input1           1.00   1.00   1.00   1.00
    wc-scanStream       0.99   1.02   1.05   1.02
    zebra               1.00   1.00   1.00   1.01
    zern                1.00   1.01   1.01   1.02

Note that MLton0 and MLton1 (ascHash) should be equivalent, from which
we conclude that run-time ratios between 0.97 and 1.03 are not
significant.  The results seem to suggest that '-cse-canon descIndex'
is the best choice.  Using '-cse-canon descHash' or '-cse-canon
ascIndex' seems to trigger a slowdown in mandelbrot.  Correspondingly,
using '-cse-canon descIndex' seems to avoid the slowdown in mandelbrot
and seems to trigger speedups in checksum, lexgen?, psdes-random.