Add `-opt-fuel <n>` expert compile-time option
A useful technique for developing optimization passes is to limit the
number of transformations made according to some finite "fuel".  The
particular transformation that triggers a bug or a performance change
can be found by binary-searching through the space of
fuels (`-opt-fuel 0` when the bug does not occur and `-opt-fuel 1000`
when the bug does occur).

A `val Control.optFuelAvailAndUse: unit -> bool` is also provided.

By default, if `-opt-fuel` is not given, then "infinite" fuel is
assumed.