Eliminate MLton_size indirection via 'a ref
The runtime GC_size function expects an object pointer (in order to
trace the reachable objects).  Previously, the Basis Library
implementation of MLton.size was of the form

   fun size x =
      let
         val refOverhead =
            Int32.toInt NormalMetaDataSize.bytes + Int.div (ObjptrWord.wordSize, 8)
      in
         C_Size.toInt (Primitive.MLton.size (ref x)) - refOverhead
      end

However, with the changes to the representation of a `unit
ref` (represented by just an object header and no content bytes), this
can result in negative sizes for objects that are represented by
`unit` (e.g., an unused function).

Now, the Basis Library implementation is of the form

   fun size x =
      C_Size.toInt (Primitive.MLton.size x)

During the translation from SSA2 to RSSA, the MLton_size primitive is
replaced by the constant 0 if the representation of the argument is
not an object pointer.  This corrects the size of objects represented
by `unit`.  It also (arguably) corrects the size of objects that could
be flattened into the `'a ref` indirection.  For example, previously,
on a 64-bit platform, the size of a `word32 * word32` could be
reported as 0, because the `(word32 * word32) ref` was flattened, so
`- refOverhead` subtracts from the actual content bytes.  Now, the
size of a `word32 * word32` is reported as 16 bytes (= 8-byte header +
4 bytes (word32) + 4 bytes (word32)).