Allow globalization of arrays
Space-safety prohibits ConstantPropagation from globalizing all arrays
that are allocated at most once by a program.  In particular, because
globals are live for the duration of the program, globalizing an `int
list array` (for example) would not be safe-for-space: arbitrarily
large lists may be stored in the array and never garbage
collected (whereas, when the `int list array` is not globalized, it
will be garbage collected when it is no longer live).

On the other hand, globalizing an `int array` (with a constant length)
is safe-for-space.

However, previously, the globalization of a `val a: t array =
Array_alloc[t] (l)` was conditioned on the smallness of `t array`.
Furthermore, `Type.isSmall` returns `false` for arrays, so no array
would be globalized.

It is correct to globalize an array if `t` is small; note that to
globalize `val a: t array = Array_alloc[t] (l)`, `l` (the length) must
be globalized and must, therefore, be a constant and the array is of
constant size. (This is Stephen Weeks's relaxed notion of
safe-for-space, where the constant factor blowup can be chosen per
program.)  In practice, it may be better to limit globalization of
arrays to ones with "small" length in addition to small element type.

This commit allows a `val a: t array = Array_alloc[t] (l)` to be
globalized if `-globalize-arrays true`, `t` is small, and `l` is
globalized.