Add `structure ConstantPropagation.Value.Raw`
Fixes a bug (triggered by the `simple` benchmark with
`-globalize-arrays true`) where the rawness of an array value was not
properly propagated.

Previously, the rawness of an array value was implemented as `bool
option ref`, but coercing `ref NONE` to `ref NONE` would not remember
the coercion when the first raw became `ref (SOME true)`.

Now, a proper "flat lattice"-like `structure Raw` is used.