Disallow globalization of IntInf and Thread types
Technically, both IntInf and Thread should not be considered small
types, as they can be of unbounded size (an IntInf can be represented
by a sequence, while a Thread implicitly contains a stack).

Note, an IntInf value can be globalized (as it is a constant), but an
`IntInf.int ref` cannot.