Fix heap invariant predicates
Previously, the `isPointerInOldGen` and `isPointerInNursery`
predicates checked that the object pointer is strictly less than
`s->heap.start + s->heap.oldGenSize` and `s->frontier`, respectively.

With MLton/mlton#219 (forward pointer in header), 0-byte objects are
allowed and the above checks are insufficient.  In particular,
immediately after allocating a 0-byte object, the object pointer is
equal to the frontier.

Now, the `isPointerInOldGen` and `isPointerInNursery` predicates are
revised to check that the end of the object is less than or equal to
`s->heap.start + s->heap.oldGenSize` and `s->frontier`, respectively.