Fix heap invariant predicates
381a47e revised the `isPointerInOldGen` and `isPointerInNursery`
predicates are revised to check that the end of the object is less
than or equal to `s->heap.start + s->heap.oldGenSize` and
`s->frontier`, respectively.  Determining the end of the object
requires using the object header (to determine the object size).

However, there are (at least) two scenarios in which the pointer given
to `isPointerInOldGen` or `isPointerInNursery` does not have a valid
header.  First, during a copying collection, `forward` asserts
`isObjptrInFromSpace (s, *opp)`; if `*opp` is already forwarded, then
the header has been overwritten with a forward pointer.  Second,
during a hash-consing dfs mark, `markIntergenerationalPointer` and
`markIntergenerationalObjptr` use `isPointerInOldGen` on interior
pointers (i.e., on a pointer to the slot in an object that has an
object pointer).

Now, the `isPointerInOldGen` and `isPointerInNursery` predicates are
revised to check that the pointer less than or equal to `s->heap.start
+ s->heap.oldGenSize` and `s->frontier`, respectively.