Fix interface of result of ElaborateEnv.functorClosure.summary
ElaborateEnv.functorClosure.summary performs a functor application
without re-elaboration of the functor body.  Using the `instantiate`
function from the `actual` dummy structure and `resultStructure` from
the initial elaboration of the functor declaration, it rewrites the
`resultStructure`, replacing any tycons from `formal` with the
corresponding tycons from `actual` and replacing any tycons generated
during the initial elaboration with fresh tycons.  However, this
replacement was not performed on the `interface` of `resultStructure`.
This could lead to anomalies in the pretty-printing of functor
closures; essentially, the tycons present in the interface were always
out of scope, because they did not correspond to the tycons from the
`actual` or returned structure.

Currently, a structure's interface is only used for
pretty-printing (to express the structure's interface in terms of a
named signature) and to short-circuit signature
matching (`ElaborateEnv.transparentCut`) if a structure is matched
against it's own signature.  Thus, the previous behavior of not
properly constructing the interface of the returned structure could
not be exploited as a type-checking bug.  [And, in retrospect, it
suggests that the short-circuiting of signature matching is never
exploited, since a structure's interface is always a realized
interface, but the to-be-matched interface is almost always a flexible
interface.]