Update pretty-printing of `signature S` and `functor F`
When pretty-printing a (flexible) interface, the interface is realized
with dummy tycons to replace flexible tycons.  We now
`setLayoutPrettyTycon` for such created dummy tycons, so that they
display as "_sig.t" rather than "?._sig.t" if `prefixUnset = true`.
Such tycons are not out of scope, so do not require a `?.` prefix.
This affects signature definitions and functor arguments.

When pretty-printing a functor closure, the result interface is
generated by applying the functor-closure's `summary` to a dummy
argument structure.  We now `loopStr` (internally using
`setLayoutPrettyTycon`) on the argument and result structure to treat
the tycons generated for the argument structure and result structure
as in scope.  The argument structure is processed with priority 1,
while the result structure is processed with priority 2, so that
tycons propagated from the argument to result are expressed in terms
of the argument structure.