Add support for `(*#showBasis "<file>"*)` directives
A comment of the form `(*#showBasis "<file>"*)` is recognized as a
directive to save the current basis to `<file>`.  The `<file>` is
interpreted relative to the source file in which it appears.  The
comment is lexed as a distinct token and is parsed as a
structure-level declaration.  [Note that treating the directive as a
top-level declaration would prohibit using it inside a functor body,
which would make the feature significantly less useful in the context
of the MLton compiler sources (with its nearly fully functorial
style).]