Flip semantics of `!` and `?` primitive operators
Previously, the `!` family of primitive operators (`+!` etc.) was an
operation that always checked for overflow, whereas the `?` family never
checked for overflow. However, as a first step towards transitioning to
a model wherein the `?` family is a predicate that checks for overflow,
here the semantics are flipped such that the `!` family never checks for
overflow, and `?` always does.