Introduce `$` operators and make `?` predicates
This change introduces a new family of `$` operators, which always check
their arguments for overflow and raise an appropriate exception. The `?`
family of operators is changed to detect overflow (returning a `Bool`),
and the `$` family is defined in terms of the corresponding `?` and `!`
operators.

Previous uses of `?` in the basis library are also corrected to use the
new system.