Make new primitives backwards-compatible
Previously the old primitives were simply deleted; this commit
introduces a compile-time switch (`MLton.newOverflow`) to switch between
the old and new systems. This should allow for easier development and
eventual integration of the new primitives, which can be incrementally
implemented on the backend while keeping the old ones intact.