Refactor checked word primitives for C codegen
A previous change introduced the use of GCC builtins for overflow
checking with the `newOverflow` flavor of checked primitive operations.
However, these builtins are compatible only with GCC versions 7 and
above, so to improve backwards compatibility a compatibility layer is
added using the older builtins for legacy GCC versions.

In addition, since much of the code for the new and old primitive
operations was shared, this shared code is pulled into the compatibility
layer (provided as `_overflow_b` macros).