Re-arrange newOverflow operation order
Previously, the new overflow-checking primitives were implemented by
first checking for overflow and then performing the operation. This
change refactors the implementation to instead perform and store the
operation first, then check for overflow.

This will allow the peephole optimization to be more easily implemented
for the new overflow primitives in the native codegens.