Perform `elimALRedundant` before `moveHoist`
Previously the `elimALRedundant` optimization was missing several cases
because it occurred after the `moveHoist` optimization. Originally, an
additional case was added to the `elimALRedundant` optimization to
handle cases where one move had been hoisted immediately before the
corresponding arithmetic operations, but this did not account for cases
where one or both moves were hoisted elsewhere in the block.

This change introduces a new group of optimizations in the
`PeepholeLivenessBlock` section which execute before `moveHoist`. In the
future additional optimizations may be added, but at present only the
`elimALRedundant` optimization occurs in this group.

In addition, the `elimALRedundant` was missing a branch to check for
imul2 operations, which has now been added.