Add `int-mul-pow2` regression test demonstrating bug
This new regression test demonstrates a bug with the native codegens and
`MLton.newOverflow true`.  Essentially, multiplications by a constant (positive
or negative) power of 2 can either erroneously `raise Overflow` or fail to
`raise Overflow`.

The issue is that the `elimMDPow2` peephole optimization tries to replace
multiplications and divisions by powers of 2 by appropriate shifts, adds, and
negations.  Unfortunately, the conditions for application of the peephole
optimization are insufficient for `MLton.newOverflow true`.

The implicit assumption in the `elimMDPow2` rewriter is that only signed
multiplication may check for overflow and that if a signed multiplication is
checked for overflow, then it does so by an immediately following iff-O
transfer.  Under these assumptions, there are special cases for multiplication
by 1, -1, 2, and -2:

    imul 1, dst       =>  jmp L_nooverflow
    jo L_overflow
    jmp L_nooverflow

    imul -1, dst      =>  neg dst
    jo L_overflow         jo L_overflow
    jmp L_nooverflow      jmp L_nooverflow

    imul 2, dst       =>  add dst, dst
    jo L_overflow         jo L_overflow
    jmp L_nooverflow      jmp L_nooverflow

    imul -2, dst      =>  add dst, dst
    jo L_overflow         neg dst
    jmp L_nooverflow      jo L_overflow
                          jmp L_nooverflow

Note that multiplication by larger powers of 2 are not rewritten, as the
resulting shift and negation will not set the overflow flag in an equivalent
manner.

If a signed multiplication is not immediately followed by an iff-condition
transfer, then multiplication by 1, -1, 2, -2, 2^i (i > 1), and -2^i (i > 1) are
rewritten:

    imul 1, dst     =>  nop

    imul -1, dst    =>  neg dst1

    imul 2, dst     =>  add dst, dst

    imul -2, dst    =>  add dst, dst
                        neg dst

    imul 2^i, dst   =>  shl i, dst

    imul -2^i, dst  =>  shl i, dst
                        neg dst

However, with `MLton.newOverflow true`, then a signed multiplication may be
followed by a `seto fdst` instruction (followed by a `test fdst, fdst`, followed
by a iff-NZ transfer --- this sequence will subsequently be optimized to an
iff-O transfer by the `conditionalJump` peephole optimization, but the
`conditionalJump` peephole optimization runs later, as part of
`peepholeLivenessBlock`).  When `imul 1, dst`, `imul 2^i, dst` (i > 1), and
`imul -2^i, dst` (i > 1) are rewritten as above, then the subsequent `seto fdst`
instruction will not be equivalent.

The appropriate fix is unclear.  One option is to strengthen the checks in the
`elimMDPow2` peephole optimization to disable the rewrite if a subsequent
`setcc` instruction would observe the flags.  Another option is to delay the
`elimMDPow2` peephole optimization until after the `elimALRedundant` peephole
optimization has executed.

As an aside, note that neither the C codegen nor the LLVM codegen have a similar
bug, although their optimizers almost certainly optimize multiplications by
powers of 2.  This simply adds to the argument that the native codegens should
be retired; resources may be better spent improving the C and LLVM codegens than
maintaining the native codegens.